Use named placeholders in CastExpression
