+++ src/Block/TryCatch/AbstractBlock.php
@@ -7,11 +7,12 @@ namespace webignition\BasilCompilableSource\Block\TryCatch;
 use webignition\BasilCompilableSource\Body\BodyInterface;
 use webignition\BasilCompilableSource\HasMetadataInterface;
 use webignition\BasilCompilableSource\Metadata\MetadataInterface;
-use webignition\Stubble\UnresolvedVariableException;
-use webignition\Stubble\VariableResolver;
+use webignition\BasilCompilableSource\RenderFromTemplateTrait;
 
 abstract class AbstractBlock implements HasMetadataInterface
 {
+    use RenderFromTemplateTrait;
+
     private BodyInterface $body;
 
     public function __construct(BodyInterface $body)
@@ -19,28 +20,9 @@ abstract class AbstractBlock implements HasMetadataInterface
         $this->body = $body;
     }
 
-    abstract protected function getRenderTemplate(): string;
-
-    /**
-     * @return array<string, string>
-     */
-    abstract protected function getRenderContext(): array;
-
     public function getMetadata(): MetadataInterface
     {
         return $this->body->getMetadata();
-    }
-
-    public function render(): string
-    {
-        try {
-            return VariableResolver::resolveTemplate(
-                $this->getRenderTemplate(),
-                $this->getRenderContext()
-            );
-        } catch (UnresolvedVariableException $unresolvedVariableException) {
-            return $unresolvedVariableException->getTemplate();
-        }
     }
 
     protected function renderBody(): string
